Final
This should be the last push, pretty sure it works fully. There were some small things that I needed to add that I didn't think of yesterday, but it works and honestly I don't wanna risk breaking it. I might make 1 more change just to the comments, but hoenstly I think that they should be fine.